管理员禁用用户、修改用户使用空间、展示文件列表

+++ wangpan_back/src/main/java/com/wangpan/dto/UserForAdminDto.java
@@ -30,8 +30,8 @@ public class UserForAdminDto {
     @JsonProperty("status")
     private Integer state;
 
-    private Integer useSpace;
+    private Long useSpace;
 
-    private Integer totalSpace;
+    private Long totalSpace;
 
 }
